Hardening quirk, stamina changes
- Adds hardening quirk
- Adds MANUAL ActivationType
- Changes to QuirkStamina
- Adds new configuration methods
- Adds player logouts for reloads
- Some class name changes

+++ src/me/simp/quirkademia/quirk/IQuirk.java
@@ -11,6 +11,6 @@ public interface IQuirk {
 	public ChatColor getChatColor();
 	public String getStaminaTitle();
 	public BarColor getStaminaColor();
-	public double getStaminaMax();
-	public double getStaminaRecharge();
+	public int getStaminaMax();
+	public int getStaminaRecharge();
 }
